[Refactor] Move JsonPathDeriver into storage layer
