+++ be/test/util/json_flattener_test.cpp
@@ -46,9 +46,9 @@
 #include "gutil/casts.h"
 #include "gutil/integral_types.h"
 #include "gutil/strings/strip.h"
+#include "storage/json_path_deriver.h"
 #include "types/json_value.h"
 #include "types/logical_type.h"
-#include "util/json_path_deriver.h"
 
 namespace starrocks {
